test: use webdriverio
Our current browser tests were incredibly flaky.

Webdriver.io seems to contain retry logic and other safeguards that we
are currently lacking in our custom browserstack-local /
selenium-webdriver setup.

Under the hood things should still work similarly, let's see if they are
now more stable.
