+++ package.json
@@ -43,9 +43,15 @@
     "@commitlint/cli": "8.3.5",
     "@commitlint/config-conventional": "8.3.4",
     "@rollup/plugin-node-resolve": "7.1.1",
+    "@wdio/browserstack-service": "5.18.7",
+    "@wdio/cli": "5.18.7",
+    "@wdio/jasmine-framework": "5.18.6",
+    "@wdio/local-runner": "5.18.7",
+    "@wdio/spec-reporter": "5.18.7",
+    "@wdio/static-server-service": "5.16.10",
+    "@wdio/sync": "5.18.7",
     "babel-eslint": "10.0.3",
     "babel-plugin-add-module-exports": "1.0.2",
-    "browserstack-local": "1.4.5",
     "bundlewatch": "0.2.5",
     "eslint": "6.8.0",
     "eslint-config-prettier": "6.10.0",
@@ -60,7 +66,6 @@
     "rollup": "1.31.1",
     "rollup-plugin-terser": "5.2.0",
     "runmd": "1.3.2",
-    "selenium-webdriver": "3.6.0",
     "standard-version": "7.1.0"
   },
   "scripts": {
@@ -73,7 +78,7 @@
     "pretest": "npm run build",
     "test": "BABEL_ENV=commonjs node --throw-deprecation node_modules/.bin/jest test/unit/",
     "pretest:browser": "npm run build && npm-run-all --parallel examples:**",
-    "test:browser": "BABEL_ENV=commonjs jest --forceExit --verbose test/browser/${BROWSER:-}*",
+    "test:browser": "wdio run ./wdio.conf.js",
     "prettier:check": "prettier --ignore-path .prettierignore --check '**/*.{js,jsx,json,md}'",
     "prettier:fix": "prettier --ignore-path .prettierignore --write '**/*.{js,jsx,json,md}'",
     "ci": "npm run lint && npm run test && npm run prettier:check && npm run docs:diff && npm run bundlewatch",
